Remove inclusion of 'parser.h' and fix pointer type

+++ modlib.h
@@ -5,8 +5,6 @@
 #include <stdlib.h>
 #include <inttypes.h>
 
-#include "parser.h"
-
 //Function prototypes
 extern uint16_t MODBUSSwapEndian( uint16_t );
-extern uint16_t MODBUSCRC16( uint16_t *, uint16_t );
+extern uint16_t MODBUSCRC16( uint8_t *, uint16_t );
